fix: Forbid deletion of current version in version expiration code
Log it instead.

+++ lib/Versions/GroupVersionsExpireManager.php
@@ -16,6 +16,7 @@ use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\EventDispatcher\IEventDispatcher;
 use OCP\Files\FileInfo;
 use OCP\IUser;
+use Psr\Log\LoggerInterface;
 
 /**
  * @psalm-import-type InternalFolderOut from FolderManager
@@ -32,7 +33,8 @@ class GroupVersionsExpireManager extends BasicEmitter {
 		ExpireManager $expireManager,
 		VersionsBackend $versionsBackend,
 		ITimeFactory $timeFactory,
-		IEventDispatcher $dispatcher
+		IEventDispatcher $dispatcher,
+		private LoggerInterface $logger,
 	) {
 		$this->folderManager = $folderManager;
 		$this->expireManager = $expireManager;
@@ -76,6 +78,18 @@ class GroupVersionsExpireManager extends BasicEmitter {
 				$versions = $this->versionsBackend->getVersionsForFile($dummyUser, $file);
 				$expireVersions = $this->expireManager->getExpiredVersion($versions, $this->timeFactory->getTime(), false);
 				foreach ($expireVersions as $version) {
+					if ($version->isCurrentVersion()) {
+						$this->logger->error(
+							'Current version of a groupfolders file was listed for deletion. Skipping.',
+							[
+								'folderid' => $version->getFolderId(),
+								'timestamp' => $version->getTimestamp(),
+								'mtime' => $version->getSourceFile()->getMtime(),
+								'sourcefilename' => $version->getSourceFileName(),
+							]
+						);
+						continue;
+					}
 					/** @var GroupVersion $version */
 					$this->emit(self::class, 'deleteVersion', [$version]);
 					$view->unlink('/' . $fileId . '/' . $version->getVersionFile()->getName());
